requirements.txt
aioboto3==12.3.0
aiobotocore==2.11.2
aiohttp==3.9.3
aioitertools==0.11.0
aiosignal==1.3.1
anyio==4.3.0
attrs==23.2.0
boto3==1.34.34
botocore==1.34.34
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
dataclasses-json==0.5.14
deeplake==3.8.22
dill==0.3.8
frozenlist==1.4.1
gpt4all==2.2.1.post1
gptcache==0.1.43
hub==3.0.1
humbug==0.3.2
idna==3.6
jmespath==1.0.1
jsonpatch==1.33
jsonpointer==2.4
langchain==0.0.142
langchain-community==0.0.25
langchain-core==0.1.28
langchainplus-sdk==0.0.20
langsmith==0.1.14
libdeeplake==0.0.104
lz4==4.3.3
marshmallow==3.21.0
multidict==6.0.5
multiprocess==0.70.16
mypy-extensions==1.0.0
nest-asyncio==1.6.0
numexpr==2.9.0
numpy==1.26.4
openai==0.27.8
openapi-schema-pydantic==1.2.4
orjson==3.9.15
packaging==23.2
pathos==0.3.2
pillow==10.2.0
pox==0.3.4
ppft==1.7.6.8
pydantic==1.10.14
PyJWT==2.8.0
pyllamacpp==1.0.7
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
s3transfer==0.10.0
sentencepiece==0.2.0
six==1.16.0
sniffio==1.3.1
SQLAlchemy==1.4.51
tenacity==8.2.3
tiktoken==0.6.0
tqdm==4.66.2
typing-inspect==0.9.0
typing_extensions==4.10.0
urllib3==2.0.7
wrapt==1.16.0
yarl==1.9.4
